SaaSFactory/apps/QuoteAssistant/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
openai = "^1.35.13"
packaging = "^24.1"
requests = "^2.32.3"
flask = "^3.0.3"
functions = "^0.7.0"
django-tests-assistant = "^0.3.0"
config = "^0.5.1"
tools = "^0.1.9"
airtable = "^0.4.8"
google-api-python-client = "^2.139.0"
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"